requirements.txt
PyMuPDF>=1.23.0
google-genai>=1.0.0
pymongo>=4.7.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
